+++ CHANGELOG.md
@@ -44,9 +44,10 @@
   - Add `filter.isSearchable`. If `false`, the filter value will be added to the url query options and not in the `search` value.
   - Every 10 minutes, the server invalidations are retrieved to compare local invalidations with server ones.
   - Add `store.isAutoRequest` atttribute to enable or disable automatic `lunaris.get` call on `reset` event when subscribing to store with `vm.stores`.
-  - Add `lunaris.collectionResultSet` to query collection data (more in documentation) !.
+  - Add `lunaris.collectionResultSet` to query collection data (more in documentation).
   - Add `dependencies` option for `lunaris.build`.
   - La valeur `''` est acceptée pour les champs définis avec l'option `optional`.
+  - Add `lunaris.dynamicView` (more in documentation).
 - Fix:
   - Improve offline app.
   - In offline or offline mode, generated urls were not encoded.